% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_cisbios}
\alias{get_cisbios}
\title{Get bios}
\usage{
get_cisbios(ibu_id)
}
\arguments{
\item{ibu_id}{Character. IBU athlete ID.
Find it, e.g. by calling \code{\link[=search_athletes]{search_athletes()}} first.}
}
\value{
A list.
}
\description{
Get a wide range of athlete-specific data.
}
\details{
Note: The list includes a wide range of information, nested in further lists
which is time-consuming to expand into one 'tidy' data.frame.
To get more specific information in tidy format, use the
functions \code{\link[=get_stats]{get_stats()}}, \code{\link[=get_recent]{get_recent()}}, and \code{\link[=get_ranks]{get_ranks()}}.
}
\examples{
get_cisbios("BTNOR11605199301")
}
