% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_competition}
\alias{get_competition}
\title{Get competitions}
\usage{
get_competition(event_id)
}
\arguments{
\item{event_id}{Character. Event ID. Find by calling \code{\link[=get_events]{get_events()}}.}
}
\value{
A data.frame.
}
\description{
Get detailed schedule of competitions
}
\examples{
get_competition(get_events("1819", "1")$EventId[1])
}
