% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_cup_results}
\alias{get_cup_results}
\title{Get cup results}
\usage{
get_cup_results(cup_id)
}
\arguments{
\item{cup_id}{Character. Cup identifier.
Find it by calling \code{\link[=get_cups]{get_cups()}}.}
}
\value{
A data.frame. Listing the standings for the queried cup.
}
\description{
Get past or current standings in a Cup.
}
\examples{
get_cup_results("BT2122SWRLCP__SWTS")
}
