% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_cups}
\alias{get_cups}
\title{Get cups}
\usage{
get_cups(season)
}
\arguments{
\item{season}{Character. Indicating the season
by the last two digits of two following years, e.g. "1819" for 2018-2019.}
}
\value{
A data.frame. Lists all cups of the queried season.
}
\description{
Get a list of all cups per season.
}
\examples{
get_cups("2122")
}
