% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_race_results}
\alias{get_race_results}
\title{Get race results}
\usage{
get_race_results(race_id)
}
\arguments{
\item{race_id}{Character. Race ID. Find by calling \code{\link[=get_competition]{get_competition()}}.}
}
\value{
A data.frame. One athlete per row. The first block of variables
contains information about the event.
}
\description{
Get the results of a race.
}
\examples{
get_race_results("BT2223SWRLCP02SMSP")
}
