% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_ranks}
\alias{get_ranks}
\title{Get ranks}
\usage{
get_ranks(ibu_id)
}
\arguments{
\item{ibu_id}{Character. IBU athlete ID.
Find it, e.g. by calling \code{\link[=search_athletes]{search_athletes()}} first.}
}
\value{
A data.frame. Lists (incomplete?) cumulative ranks of the athlete.
The variable 'Description' indicates the ranks.
}
\description{
Get the athlete's cumulative ranks
}
\details{
Note: Due to lacking API documentation, it is somewhat unclear what is returned here.
If you need a complete list of all results of that athlete, use the call \code{\link[=get_results]{get_results()}}.
}
\examples{
get_ranks("BTNOR11605199301")
}
