% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_recent}
\alias{get_recent}
\title{Get recent results}
\usage{
get_recent(ibu_id)
}
\arguments{
\item{ibu_id}{Character. IBU athlete ID.
Find it, e.g. by calling \code{\link[=search_athletes]{search_athletes()}} first.}
}
\value{
A data.frame. Lists recent race results.
}
\description{
Get the athlete's recent performaces in races.
}
\examples{
get_recent("BTNOR11605199301")
}
