% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_results}
\alias{get_results}
\title{Get results per athlete}
\usage{
get_results(ibu_id)
}
\arguments{
\item{ibu_id}{Character. IBU athlete ID.
Find it, e.g. by calling \code{\link[=search_athletes]{search_athletes()}} first.}
}
\value{
A data.frame. Lists all race results of an athlete.
}
\description{
Get results per athlete
}
\examples{
get_results("BTNOR11605199301")
}
