% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_stats}
\alias{get_stats}
\title{Get stats}
\usage{
get_stats(ibu_id)
}
\arguments{
\item{ibu_id}{Character. IBU athlete ID.
Find it, e.g. by calling \code{\link[=search_athletes]{search_athletes()}} first.}
}
\value{
A data.frame. Lists overall shooting and skiing performance stats.
}
\description{
Get the athlete's shooting and skiing stats for the last four seasons.
}
\examples{
get_stats("BTNOR11605199301")
}
